+++ app/Models/EventLog.php
@@ -31,7 +31,6 @@ use Adshares\Adserver\Utilities\DomainReader;
 use Adshares\Common\Domain\ValueObject\Uuid;
 use Adshares\Supply\Application\Dto\UserContext;
 use DateTime;
-use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
@@ -40,6 +39,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use function hex2bin;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Collection as SupportCollection;
 use stdClass;
 
 /**
@@ -205,14 +205,6 @@ SQL;
         return $result;
     }
 
-    public static function fetchEvents(Arrayable $paymentIds, int $limit, int $offset): Collection
-    {
-        return self::whereIn('payment_id', $paymentIds)
-            ->limit($limit)
-            ->offset($offset)
-            ->get();
-    }
-
     public static function create(
         string $caseId,
         string $eventId,
@@ -348,6 +340,15 @@ SQL;
         return self::whereIn('event_id', $binEventIds)->get();
     }
 
+    public static function fetchCreationHourTimestampByIds(array $ids): array
+    {
+        return self::select(DB::raw('DISTINCT FLOOR(UNIX_TIMESTAMP(created_at)/3600)*3600 AS ts'))
+            ->whereIn('id', $ids)
+            ->get()
+            ->pluck('ts')
+            ->all();
+    }
+
     public static function fetchLastByTrackingId(string $campaignPublicId, string $trackingId): ?self
     {
         return self::where('campaign_id', hex2bin($campaignPublicId))
@@ -382,19 +383,13 @@ SQL;
         $this->our_userdata = $userContext->keywords();
     }
 
-    public function setStatus(int $status): void
+    public static function updatePaymentIdByIds(int $paymentId, SupportCollection $ids): int
     {
-        $this->payment_status = $status;
-        $this->save();
-    }
-
-    public function setValueAndStatus(int $valueCurrency, float $exchangeRateValue, int $value, int $status): void
-    {
-        $this->event_value_currency = $valueCurrency;
-        $this->exchange_rate = $exchangeRateValue;
-        $this->event_value = $value;
-        $this->payment_status = $status;
-        $this->save();
+        return self::whereIn('id', $ids)
+            ->update([
+                'payment_id' => $paymentId,
+                'updated_at' => new DateTime(),
+            ]);
     }
 
     public function conversions(): HasMany